Made fixes to slack models
