+++ src/main/java/com/blackducksoftware/integration/hub/alert/datasource/entity/distribution/SlackDistributionConfigEntity.java
@@ -36,16 +36,12 @@ public class SlackDistributionConfigEntity extends DatabaseEntity {
     @Column(name = "channel_name")
     private String channelName;
 
-    @Column(name = "channel_username")
-    private String channelUsername = "Hub-alert";
-
     public SlackDistributionConfigEntity() {
     }
 
-    public SlackDistributionConfigEntity(final String channelName, final String channelUsername) {
+    public SlackDistributionConfigEntity(final String channelName) {
         super();
         this.channelName = channelName;
-        this.channelUsername = channelUsername;
     }
 
     public static long getSerialversionuid() {
@@ -54,10 +50,6 @@ public class SlackDistributionConfigEntity extends DatabaseEntity {
 
     public String getChannelName() {
         return channelName;
-    }
-
-    public String getChannelUsername() {
-        return channelUsername;
     }
 
 }
